+++ HISTORY.rst
@@ -7,7 +7,8 @@
   handled by the provider.
 * Let provider handle superordinates and subordinate arrays.
 * Let provider add notes to collections.
-
+* Added a :class:`~skosprovider_sqlalchemy.models.Match` model to handle
+  matches.
 
 0.2.1 (2014-08-25)
 ------------------